requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
notebook
xgboost
PyYAML
optuna
streamlit